refactor: move cache files to search folder
